(fix) Fix error due to incomplete rename
